pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.isort]
profile = "black"

[tool.poetry]
authors = ["The Palace Project <info@thepalaceproject.org>"]
description = "Virtual Library Card Creator"
homepage = "https://thepalaceproject.org"
name = "Virtual Library Card"
readme = "README.md"
repository = "https://github.com/ThePalaceProject/ virtual-library-card"
version = "0" # Version number is managed with tags in git

[tool.poetry.dependencies]
attrs = "^22.1.0"
boto3 = "^1.26.22"
certifi = "2022.9.24"
chardet = "^5.1.0"
crispy-forms-materialize = "0.2"
datedelta = "1.4"
Django = "3.2.16"
django-absoluteuri = "1.3.0"
django-appconf = "1.0.5"
django-bootstrap4 = "2.3.1"
django-compressor = "2.4"
django-compressor-postcss = "0.8.2"
django-crispy-forms = "1.14.0"
django-htmlmin = "0.11.0"
django-localflavor = "3.1"
django-localflavor-us = "1.1"
django-material = "1.7.4"
django-materialize-css = "0.0.1"
django-recaptcha = "^3.0.0"
django-rename-app = "^0.1.5"
django-render-block = "0.9.2"
django-sequences = "2.5"
django-storages = "^1.13.1"
djangorestframework = "3.14.0"
jinja2 = "3.1.2"
libsass = "0.20.1"
Pillow = "9.3.0"
psycopg2-binary = "2.9.5"
PyJWT = "^2.6.0"
PyMySQL = "1.0.2"
python = ">=3.7,<4"
python-stdnum = "1.15"
pytz = "2022.6"
rcssmin = "1.0.6"
requests = "2.28.1"
rjsmin = "1.1.0"
six = "1.16.0"
smartystreets-python-sdk = "4.11.9"
uwsgi = ">=2.0,<2.1"

[tool.poetry.group.ci.dependencies]
pre-commit = "~2.17"
tox = "^3.26.0"
tox-docker = "^3.1.0"
tox-gh-actions = "^2.10.0"

[tool.poetry.group.dev.dependencies]
mock = { version = "~4.0", python = "< 3.8" }
parameterized = "^0.8.1"
pytest = "^7.1.1"
pytest-django = "^4.5.2"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "virtual_library_card.settings.dev"
